Fix warning: implicit conversion changes signedness
visp/tutorial/segmentation/color/tutorial-hsv-segmentation.cpp:127:15: warning: implicit conversion changes signedness: 'unsigned int' to 'int' [-Wsign-conversion]
  127 |     width = I.getWidth();
      |           ~ ~~^~~~~~~~~~
visp/tutorial/segmentation/color/tutorial-hsv-segmentation.cpp:128:16: warning: implicit conversion changes signedness: 'unsigned int' to 'int' [-Wsign-conversion]
  128 |     height = I.getHeight();
      |            ~ ~~^~~~~~~~~~~
visp/tutorial/segmentation/color/tutorial-hsv-segmentation.cpp:131:39: warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
  131 |   vpImage<unsigned char> mask(height, width);
      |                          ~~~~         ^~~~~
visp/tutorial/segmentation/color/tutorial-hsv-segmentation.cpp:131:31: warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
  131 |   vpImage<unsigned char> mask(height, width);
      |                          ~~~~ ^~~~~~
visp/tutorial/segmentation/color/tutorial-hsv-segmentation.cpp:132:39: warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
  132 |   vpImage<vpRGBa> I_segmented(height, width);
      |                   ~~~~~~~~~~~         ^~~~~
visp/tutorial/segmentation/color/tutorial-hsv-segmentation.cpp:132:31: warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
  132 |   vpImage<vpRGBa> I_segmented(height, width);
      |                   ~~~~~~~~~~~ ^~~~~~
visp/tutorial/segmentation/color/tutorial-hsv-segmentation.cpp:137:103: warning: implicit conversion changes signedness: 'unsigned int' to 'int' [-Wsign-conversion]
  137 |   std::shared_ptr<vpDisplay> d_I_segmented = vpDisplayFactory::createDisplay(I_segmented, I.getWidth()+75, 0, "HSV segmented frame");
      |                                              ~~~~~~~~~~~~~~~~                             ~~~~~~~~~~~~^~~

+++ tutorial/segmentation/color/tutorial-hsv-range-tuner.cpp
@@ -217,7 +217,7 @@ int main(int argc, const char *argv[])
   vpImage<unsigned char> S(height, width);
   vpImage<unsigned char> V(height, width);
   vpDisplay *d_I = vpDisplayFactory::allocateDisplay(I, 0, 0, "Current frame");
-  vpDisplay *d_I_segmented = vpDisplayFactory::allocateDisplay(I_segmented, I.getWidth()+75, 0, "Segmented frame");
+  vpDisplay *d_I_segmented = vpDisplayFactory::allocateDisplay(I_segmented, static_cast<int>(I.getWidth())+75, 0, "Segmented frame");
 #endif
   bool quit = false;
